+++ curves/src/bezier/mod.rs
@@ -52,7 +52,7 @@ pub trait BezierCurve: Clone+Sized {
     /// Given a value t from 0 to 1, returns a point on this curve
     /// 
     #[inline]
-    fn point_at_pos(&self, t: f32) -> Self::Point {
+    fn point_at_pos(&self, t: f64) -> Self::Point {
         let control_points = self.control_points();
         de_casteljau4(t, self.start_point(), control_points.0, control_points.1, self.end_point())
     }
@@ -61,7 +61,7 @@ pub trait BezierCurve: Clone+Sized {
     /// Given a value t from 0 to 1, finds a point on this curve and subdivides it, returning the two resulting curves
     /// 
     #[inline]
-    fn subdivide(&self, t: f32) -> (Self, Self) {
+    fn subdivide(&self, t: f64) -> (Self, Self) {
         let control_points              = self.control_points();
         let (first_curve, second_curve) = subdivide4(t, self.start_point(), control_points.0, control_points.1, self.end_point());
 
@@ -85,7 +85,7 @@ pub trait BezierCurve: Clone+Sized {
     /// Finds the t values where this curve has extremities
     /// 
     #[inline]
-    fn find_extremities(&self) -> Vec<f32> {
+    fn find_extremities(&self) -> Vec<f64> {
         let start       = self.start_point();
         let end         = self.end_point();
         let (cp1, cp2)  = self.control_points();
@@ -97,19 +97,19 @@ pub trait BezierCurve: Clone+Sized {
     /// Generates a curve by attempting to find a best fit against a set of points
     /// 
     #[inline]
-    fn fit_from_points(points: &[Self::Point], max_error: f32) -> Option<Vec<Self>> {
+    fn fit_from_points(points: &[Self::Point], max_error: f64) -> Option<Vec<Self>> {
         fit_curve(points, max_error)
     }
 
     ///
     /// Attempts to estimate the length of this curve
     /// 
-    fn estimate_length(&self, max_t: f32) -> f32 {
+    fn estimate_length(&self, max_t: f64) -> f64 {
         let mut last_pos = self.point_at_pos(0.0);
         let mut length   = 0.0;
 
         for t in 1..LENGTH_SUBDIVISIONS {
-            let t           = (t as f32) / (LENGTH_SUBDIVISIONS as f32) * max_t;
+            let t           = (t as f64) / (LENGTH_SUBDIVISIONS as f64) * max_t;
             let next_pos    = self.point_at_pos(t);
 
             length += last_pos.distance_to(&next_pos);
