+++ curves/src/bezier/offset.rs
@@ -3,7 +3,7 @@ use super::*;
 ///
 /// Computes a series of curves that approximate an offset curve from the specified origin curve
 /// 
-pub fn offset<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(curve: &Curve, initial_offset: f32, final_offset: f32) -> Vec<Curve> {
+pub fn offset<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(curve: &Curve, initial_offset: f64, final_offset: f64) -> Vec<Curve> {
     // Pass through the curve if it's 0-length
     let start       = curve.start_point();
     let end         = curve.end_point();
@@ -40,7 +40,7 @@ pub fn offset<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point
 /// Splits a curve at a given set of ordered offsets, returning a list of curves and
 /// their final offsets
 /// 
-fn split_offsets<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(curve: &Curve, initial_offset: f32, final_offset: f32, split_points: &[f32]) -> Vec<(Curve, f32)> {
+fn split_offsets<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(curve: &Curve, initial_offset: f64, final_offset: f64, split_points: &[f64]) -> Vec<(Curve, f64)> {
     let mut curves_and_offsets  = vec![];
     let mut remaining           = curve.clone();
     let mut remaining_t         = 0.0;
@@ -79,7 +79,7 @@ fn split_offsets<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Po
 /// Computes the offset error between a curve and a proposed offset curve at a given t value
 /// 
 #[inline]
-fn offset_error<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(original_curve: &Curve, offset_curve: &Curve, t: f32, initial_offset: f32, final_offset: f32) -> Point {
+fn offset_error<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(original_curve: &Curve, offset_curve: &Curve, t: f64, initial_offset: f64, final_offset: f64) -> Point {
     // Work out how much we need to offset the mid-point
     let midpoint_offset     = (final_offset - initial_offset) * (original_curve.estimate_length(t)/original_curve.estimate_length(1.0)) + initial_offset;
     let midpoint_normal     = original_curve.normal_at_pos(t).to_unit_vector();
@@ -96,7 +96,7 @@ fn offset_error<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Poi
 /// 
 /// This won't produce an accurate offset if the curve doubles back on itself. The return value is the curve and the error
 /// 
-fn simple_offset<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(curve: Curve, initial_offset: f32, final_offset: f32) -> (Curve, f32) {
+fn simple_offset<Point: Coordinate+Normalize<Point>, Curve: BezierCurve<Point=Point>+NormalCurve<Curve>>(curve: Curve, initial_offset: f64, final_offset: f64) -> (Curve, f64) {
     // Fetch the original points
     let start       = curve.start_point();
     let end         = curve.end_point();
