+++ curves/src/coordinate.rs
@@ -3,11 +3,11 @@ use std::ops::*;
 ///
 /// Represents a value that can be used as a coordinate in a bezier curve
 /// 
-pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f32, Output=Self>+Sub<Self, Output=Self> {
+pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f64, Output=Self>+Sub<Self, Output=Self> {
     ///
     /// Creates a new coordinate from the specified set of components
     /// 
-    fn from_components(components: &[f32]) -> Self;
+    fn from_components(components: &[f64]) -> Self;
 
     ///
     /// Returns the origin coordinate
@@ -22,7 +22,7 @@ pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f32, Output=Self>+S
     ///
     /// Retrieves the component at the specified index
     /// 
-    fn get(&self, index: usize) -> f32;
+    fn get(&self, index: usize) -> f64;
 
     ///
     /// Returns a point made up of the biggest components of the two points
@@ -38,18 +38,18 @@ pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f32, Output=Self>+S
     /// Computes the distance between this coordinate and another of the same type
     /// 
     #[inline]
-    fn distance_to(&self, target: &Self) -> f32 {
+    fn distance_to(&self, target: &Self) -> f64 {
         let offset              = *self - *target;
         let squared_distance    = offset.dot(&offset);
 
-        f32::sqrt(squared_distance)
+        f64::sqrt(squared_distance)
     }
 
     ///
     /// Computes the dot product for this vector along with another vector
     /// 
     #[inline]
-    fn dot(&self, target: &Self) -> f32 {
+    fn dot(&self, target: &Self) -> f64 {
         let mut dot_product = 0.0;
 
         for component_index in 0..Self::len() {
@@ -63,8 +63,8 @@ pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f32, Output=Self>+S
     /// Computes the magnitude of this vector
     /// 
     #[inline]
-    fn magnitude(&self) -> f32 {
-        f32::sqrt(self.dot(self))
+    fn magnitude(&self) -> f64 {
+        f64::sqrt(self.dot(self))
     }
 
     ///
@@ -99,7 +99,7 @@ pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f32, Output=Self>+S
     /// adjust each point according to its neighbours (the central weight is what's
     /// applied to the 'current' point)
     /// 
-    fn smooth(points: &[Self], weights: &[f32]) -> Vec<Self> {
+    fn smooth(points: &[Self], weights: &[f64]) -> Vec<Self> {
         let mut smoothed    = vec![];
         let points_len      = points.len() as i32;
         let weight_len      = weights.len() as i32;
@@ -135,30 +135,30 @@ pub trait Coordinate : Sized+Copy+Add<Self, Output=Self>+Mul<f32, Output=Self>+S
 /// Represents a coordinate with a 2D position
 /// 
 pub trait Coordinate2D {
-    fn x(&self) -> f32;
-    fn y(&self) -> f32;
+    fn x(&self) -> f64;
+    fn y(&self) -> f64;
 }
 
 ///
 /// Represents a coordinate with a 3D position
 /// 
 pub trait Coordinate3D {
-    fn x(&self) -> f32;
-    fn y(&self) -> f32;
-    fn z(&self) -> f32;
+    fn x(&self) -> f64;
+    fn y(&self) -> f64;
+    fn z(&self) -> f64;
 }
 
-impl Coordinate for f32 {
-    fn from_components(components: &[f32]) -> f32 {
+impl Coordinate for f64 {
+    fn from_components(components: &[f64]) -> f64 {
         components[0]
     }
 
-    #[inline] fn origin() -> f32 { 0.0 }
+    #[inline] fn origin() -> f64 { 0.0 }
     #[inline] fn len() -> usize { 1 }
-    #[inline] fn get(&self, _index: usize) -> f32 { *self }
+    #[inline] fn get(&self, _index: usize) -> f64 { *self }
 
     #[inline]
-    fn from_biggest_components(p1: f32, p2: f32) -> f32 {
+    fn from_biggest_components(p1: f64, p2: f64) -> f64 {
         if p1 > p2 {
             p1
         } else {
@@ -167,7 +167,7 @@ impl Coordinate for f32 {
     }
 
     #[inline]
-    fn from_smallest_components(p1: f32, p2: f32) -> f32 {
+    fn from_smallest_components(p1: f64, p2: f64) -> f64 {
         if p1 < p2 {
             p1
         } else {
@@ -176,25 +176,25 @@ impl Coordinate for f32 {
     }
 
     #[inline]
-    fn distance_to(&self, target: &f32) -> f32 {
-        f32::abs(self-target)
+    fn distance_to(&self, target: &f64) -> f64 {
+        f64::abs(self-target)
     }
 
-    fn dot(&self, target: &f32) -> f32 {
+    fn dot(&self, target: &f64) -> f64 {
         self * target
     }
 }
 
 /// Represents a 2D point
 #[derive(Copy, Clone, PartialEq, Debug)]
-pub struct Coord2(pub f32, pub f32);
+pub struct Coord2(pub f64, pub f64);
 
 impl Coordinate2D for Coord2 {
     ///
     /// X component of this coordinate
     /// 
     #[inline]
-    fn x(&self) -> f32 {
+    fn x(&self) -> f64 {
         self.0
     }
 
@@ -202,7 +202,7 @@ impl Coordinate2D for Coord2 {
     /// Y component of this coordinate
     /// 
     #[inline]
-    fn y(&self) -> f32 {
+    fn y(&self) -> f64 {
         self.1
     }
 }
@@ -225,18 +225,18 @@ impl Sub<Coord2> for Coord2 {
     }
 }
 
-impl Mul<f32> for Coord2 {
+impl Mul<f64> for Coord2 {
     type Output=Coord2;
 
     #[inline]
-    fn mul(self, rhs: f32) -> Coord2 {
+    fn mul(self, rhs: f64) -> Coord2 {
         Coord2(self.0 * rhs, self.1 * rhs)
     }
 }
 
 impl Coordinate for Coord2 {
     #[inline]
-    fn from_components(components: &[f32]) -> Coord2 {
+    fn from_components(components: &[f64]) -> Coord2 {
         Coord2(components[0], components[1])
     }
 
@@ -249,7 +249,7 @@ impl Coordinate for Coord2 {
     fn len() -> usize { 2 }
 
     #[inline]
-    fn get(&self, index: usize) -> f32 { 
+    fn get(&self, index: usize) -> f64 { 
         match index {
             0 => self.0,
             1 => self.1,
@@ -258,23 +258,23 @@ impl Coordinate for Coord2 {
     }
 
     fn from_biggest_components(p1: Coord2, p2: Coord2) -> Coord2 {
-        Coord2(f32::from_biggest_components(p1.0, p2.0), f32::from_biggest_components(p1.1, p2.1))
+        Coord2(f64::from_biggest_components(p1.0, p2.0), f64::from_biggest_components(p1.1, p2.1))
     }
 
     fn from_smallest_components(p1: Coord2, p2: Coord2) -> Coord2 {
-        Coord2(f32::from_smallest_components(p1.0, p2.0), f32::from_smallest_components(p1.1, p2.1))
+        Coord2(f64::from_smallest_components(p1.0, p2.0), f64::from_smallest_components(p1.1, p2.1))
     }
 
     #[inline]
-    fn distance_to(&self, target: &Coord2) -> f32 {
+    fn distance_to(&self, target: &Coord2) -> f64 {
         let dist_x = target.0-self.0;
         let dist_y = target.1-self.1;
 
-        f32::sqrt(dist_x*dist_x + dist_y*dist_y)
+        f64::sqrt(dist_x*dist_x + dist_y*dist_y)
     }
 
     #[inline]
-    fn dot(&self, target: &Self) -> f32 {
+    fn dot(&self, target: &Self) -> f64 {
         self.0*target.0 + self.1*target.1
     }
 }
