+++ curves/tests/bezier/mod.rs
@@ -9,8 +9,8 @@ mod normal;
 mod bounds;
 mod deform;
 
-pub fn approx_equal(a: f32, b: f32) -> bool {
-    f32::floor(f32::abs(a-b)*10000.0) == 0.0
+pub fn approx_equal(a: f64, b: f64) -> bool {
+    f64::floor(f64::abs(a-b)*10000.0) == 0.0
 }
 
 #[test]
@@ -27,7 +27,7 @@ fn can_read_curve_points() {
     let curve = bezier::Curve::from_points(Coord2(1.0, 1.0), Coord2(2.0, 2.0), Coord2(3.0, 3.0), Coord2(4.0, 4.0));
 
     for x in 0..100 {
-        let t = (x as f32)/100.0;
+        let t = (x as f64)/100.0;
 
         let point           = curve.point_at_pos(t);
         let another_point   = bezier::de_casteljau4(t, Coord2(1.0, 1.0), Coord2(3.0, 3.0), Coord2(4.0, 4.0), Coord2(2.0, 2.0));
