+++ curves/tests/coordinates.rs
@@ -12,7 +12,7 @@ fn can_find_unit_vector() {
     assert!(Coord2(0.0, 1.0).to_unit_vector() == Coord2(0.0, 1.0));
     assert!(Coord2(0.0, 2.0).to_unit_vector() == Coord2(0.0, 1.0));
 
-    assert!(f32::abs(Coord2(4.0, 2.0).to_unit_vector().distance_to(&Coord2(0.0, 0.0))-1.0) < 0.01);
+    assert!(f64::abs(Coord2(4.0, 2.0).to_unit_vector().distance_to(&Coord2(0.0, 0.0))-1.0) < 0.01);
 }
 
 #[test]
